requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.110.2
flatbuffers==24.3.25
fonttools==4.51.0
gast==0.5.4
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.2
h11==0.14.0
h5py==3.11.0
httplib2==0.22.0
idna==3.7
imutils==0.5.4
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
joblib==1.4.0
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==3.3.2
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pyparsing==3.1.2
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.0.2
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
scikit-learn==1.4.2
scipy==1.13.0
selenium==4.20.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.37.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
threadpoolctl==3.4.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
wcwidth==0.2.13
webdriver-manager==4.0.1
Werkzeug==3.0.2
wrapt==1.16.0
wsproto==1.2.0
